Restructure files + increase tests coverage
