+++ tests/unit/conftest.py
@@ -19,6 +19,7 @@ from unittest.mock import MagicMock, patch
 import neo4j
 import pytest
 from neo4j_graphrag.embeddings.base import Embedder
+from neo4j_graphrag.experimental.pipeline import Component
 from neo4j_graphrag.llm import LLMInterface
 from neo4j_graphrag.retrievers import (
     HybridRetriever,
@@ -98,3 +99,8 @@ def result_formatter() -> Callable[[neo4j.Record], RetrieverResultItem]:
         )
 
     return format_function
+
+
+@pytest.fixture(scope="function")
+def component() -> MagicMock:
+    return MagicMock(spec=Component)
